Add converted agent service
